requirements.txt
absl-py==0.3.0
appdirs==1.4.3
appnope==0.1.0
asn1crypto==0.24.0
astor==0.7.1
astroid==1.6.5
attrs==18.1.0
Automat==0.7.0
autopep8==1.3.5
backcall==0.1.0
bleach==2.1.3
certifi==2018.4.16
cffi==1.11.5
cloudpickle==0.5.3
constantly==15.1.0
cryptography==2.2.2
cycler==0.10.0
dask==0.18.2
decorator==4.3.0
entrypoints==0.2.3
gast==0.2.0
grpcio==1.13.0
h5py==2.8.0
html5lib==1.0.1
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
ipykernel==4.8.2
ipython==6.4.0
ipython-genutils==0.2.0
ipywidgets==7.3.0
isort==4.3.4
jedi==0.12.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
mccabe==0.6.1
mistune==0.8.3
mkl-fft==1.0.2
mkl-random==1.0.1
mpmath==1.0.0
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
nose==1.3.7
notebook==5.6.0
numpy==1.15.0
olefile==0.45.1
pandas==0.23.3
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.1.0
progress==1.4
prometheus-client==0.3.0
prompt-toolkit==1.0.15
protobuf==3.6.0
ptyprocess==0.6.0
pyasn1==0.4.3
pyasn1-modules==0.2.2
pycodestyle==2.4.0
pycparser==2.18
Pygments==2.2.0
pylint==1.9.2
pyOpenSSL==18.0.0
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.5
PyWavelets==0.5.2
pyzmq==17.1.0
qtconsole==4.3.1
scikit-image==0.14.0
scikit-learn==0.19.2
scipy==1.1.0
Send2Trash==1.5.0
service-identity==17.0.0
simplegeneric==0.8.1
six==1.11.0
sympy==1.2
tensorboard==1.9.0
tensorboardX==1.2
tensorflow==1.9.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
toolz==0.9.0
torch==0.4.0
torchvision==0.2.1
tornado==5.1
traitlets==4.3.2
Twisted==17.5.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.3.0
wrapt==1.10.11
zope.interface==4.5.0
